+++ packages/app/src/pages/api/data/timeseries/[...param].ts
@@ -11,27 +11,11 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import path from 'path';
 import sanitize from 'sanitize-filename';
 import { isDefined } from 'ts-is-present';
+import { resolvePublicFolder } from '~/utils/api/resolve-public-folder';
 import { countTrailingNullValues } from '~/utils/count-trailing-null-values';
 
 const publicPath = resolvePublicFolder(path.resolve(__dirname));
 const publicJsonPath = path.resolve(publicPath, 'json');
-
-function resolvePublicFolder(cwd: string): string {
-  const parentPath = path.resolve(cwd, '..');
-
-  //This happens when we've navigated all the way up to the root and can't go any higher
-  if (parentPath === cwd) {
-    throw new Error('Unable to resolve public folder');
-  }
-
-  const publicPath = path.join(parentPath, 'public');
-
-  if (fs.existsSync(publicPath)) {
-    return publicPath;
-  }
-
-  return resolvePublicFolder(parentPath);
-}
 
 export default function handler(req: NextApiRequest, res: NextApiResponse) {
   if (!Array.isArray(req.query.param)) {
